% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_pkg_sbom.R
\name{cran_pkg_sbom}
\alias{cran_pkg_sbom}
\title{Create a minimal SBOM for a CRAN R package}
\usage{
cran_pkg_sbom(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
JSON SBOM string
}
\description{
Create a minimal SBOM for a CRAN R package
}
