% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbom-package.R
\docType{package}
\name{sbom}
\alias{sbom}
\alias{sbom-package}
\title{Generate Software Bill of Materials for R Things}
\description{
A "Software Bill of Materials" (SBOM) is a nested inventory for software,
a list of ingredients that make up software components. Tools are provided
to make SBOMs for R things.
}
\references{
\url{https://www.ntia.gov/SBOM}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/sbom}
  \item Report bugs at \url{https://github.com/hrbrmstr/sbom/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
