% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-sbom.R
\name{validate_sbom}
\alias{validate_sbom}
\title{Validate an SBOM using CycloneDX}
\usage{
validate_sbom(path = NULL, text = NULL)
}
\arguments{
\item{path}{path to a CycloneDX SBOM JSON file or \code{NULL} if using \code{text}. Will be
\code{\link[=path.expand]{path.expand()}}ed.}

\item{text}{CycloneDX SBOM JSON string or \code{NULL} if using \code{path}}
}
\description{
If the \code{cyclonedx} binary is available, use it to valiate an SBOM.
}
