% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbom-package.R
\docType{package}
\name{sbom}
\alias{sbom}
\alias{sbom-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/sbom}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/sbom/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
